Fix RREP emission on second successor selection
When 2 successors was used, the first one was aware of the node host
route; however, the node did not send a RREP through the second
successor as it hasn't moved in the tree (it was always reachable
through the previous successor). Then, the second successors did not
know it as predecessor, and did not route its packets.