Add docker container creation scripts
`docker_image/` contains all needed stuff to run a docker container,
for testing LRP. To run LRP with its own network stack, I use this
command:

    docker run --network "$ISOLATED_NETWORK_NAME" \
               --volume "${PROJECT_ROOT}:/root/pylrp" \
               --privileged \
               --name "$CONTAINER_NAME" --hostname "$CONTAINER_NAME" \
               --entrypoint '/root/pylrp/docker_image/init_lrp.sh' \
               --workdir /root/pylrp/src \
               --env LANG=fr_FR.UTF-8 \
               --interactive --tty \
               "${IMAGE_NAME}" \
               python -m lrp daemon eth0

Optionnally, for the sink, add `0 --sink` to the last line.

To have a bash prompt on top of this node, e.g. to be able to see the
routing table / to ping other machines, use:

    docker exec -it "$CONTAINER_NAME" /bin/bash